% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.mle.R
\name{ratioTest}
\alias{ratioTest}
\title{Internal ratio test helper}
\usage{
ratioTest(dat, whichTau, dist = "exp")
}
\arguments{
\item{dat}{A data frame with columns \code{Y} and \code{D}.}

\item{whichTau}{Evaluation time point.}

\item{dist}{Distribution name.}
}
\value{
A numeric vector of estimates.
}
\description{
Internal ratio test helper
}
\keyword{internal}
