% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cure.model.fit.R
\name{plot.cure.model.fit}
\alias{plot.cure.model.fit}
\title{Plot a fitted cure model object}
\usage{
\method{plot}{cure.model.fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cure.model.fit"}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Displays the Kaplan–Meier plot stored in a \code{cure.model.fit} object.
}
