% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.surv.data.R
\name{prepare.surv.data}
\alias{prepare.surv.data}
\title{Prepare survival data for cure model assessment}
\usage{
prepare.surv.data(data, time, status, time_scale = c("none", "days_to_years"))
}
\arguments{
\item{data}{A data frame.}

\item{time}{Character string giving the survival time column name.}

\item{status}{Character string giving the event indicator column name.}

\item{time_scale}{Either \code{"none"} or \code{"days_to_years"}.}
}
\value{
A data frame with standardized columns \code{Y} for survival time
and \code{D} for event indicator, D = 1 for event and D = 0 for censoring.
}
\description{
Standardizes survival data into a format required by the package.
The returned data frame always contains columns \code{Y} for survival time
and \code{D} for event indicator, D = 1 for event and D = 0 for censoring.
}
