% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cure.model.fit.R
\name{summary.cure.model.fit}
\alias{summary.cure.model.fit}
\title{Summarize a fitted cure model object}
\usage{
\method{summary}{cure.model.fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cure.model.fit"}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A data frame containing the AIC comparison of candidate models.
}
\description{
Returns the AIC comparison table from a \code{cure.model.fit} object.
}
\examples{
library(survival)

dat <- prepare.surv.data(gbsg, "rfstime", "status", "days_to_years")
fit_res <- model.fitting(dat, plot_km = FALSE)

summary(fit_res)

}
